+++ yaqd_control/_status.py
@@ -10,23 +10,28 @@ from rich.console import Console
 
 import yaqc  # type: ignore
 
-from ._cache import read_daemon_cache
+from ._cache import read_daemon_cache, write_to_daemon_cache
 
 
 def connect(daemon):
     c = yaqc.Client(host=daemon.host, port=daemon.port)
     busy = c.busy()
-    return busy
+    name = c.id()["name"]
+    return busy, name
 
 
-def fill(busy, online_text, busy_text):
+def fill(result, online_text, busy_text, name_text, cached_daemon):
+    busy, name = result
     online_text.append("online", style="green")
     busy_text.append(str(busy), style="red" if busy else "green")
+    name_text.append(name)
+    cached_daemon.name = name
 
 
-def fill_error(e, online_text, busy_text):
+def fill_error(e, online_text, busy_text, name_text, cached_daemon):
     online_text.append("offline", style="red")
     busy_text.append("?", style="red")
+    name_text.append(cached_daemon.name)
 
 
 def status(force_color=False):
@@ -41,14 +46,16 @@ def status(force_color=False):
         console = None
     with Live(table, refresh_per_second=4, console=console) as live:
         results = []
-        for daemon in read_daemon_cache():
+        daemons = read_daemon_cache()
+        for daemon in daemons:
             online_text = Text("")
             busy_text = Text("")
+            name_text = Text("")
             table.add_row(
                 daemon.host,
                 str(daemon.port),
                 daemon.kind,
-                daemon.name,
+                name_text,
                 online_text,
                 busy_text,
             )
@@ -56,9 +63,19 @@ def status(force_color=False):
                 pool.apply_async(
                     connect,
                     (daemon,),
-                    callback=partial(fill, online_text=online_text, busy_text=busy_text),
+                    callback=partial(
+                        fill,
+                        online_text=online_text,
+                        busy_text=busy_text,
+                        name_text=name_text,
+                        cached_daemon=daemon,
+                    ),
                     error_callback=partial(
-                        fill_error, online_text=online_text, busy_text=busy_text
+                        fill_error,
+                        online_text=online_text,
+                        busy_text=busy_text,
+                        name_text=name_text,
+                        cached_daemon=daemon,
                     ),
                 )
             )
@@ -66,3 +83,6 @@ def status(force_color=False):
         # Wait for all the results before exiting live view
         for r in results:
             r.wait()
+
+    # update cache (names may have changed)
+    write_to_daemon_cache(*[d for d in daemons])